% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{AnalyticsMetricsBase}
\alias{AnalyticsMetricsBase}
\title{AnalyticsMetricsBase Class}
\description{
Base class for tuning job or training job analytics classes. Understands
             common functionality like persistence and caching.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnalyticsMetricsBase-new}{\code{AnalyticsMetricsBase$new()}}
\item \href{#method-AnalyticsMetricsBase-export_csv}{\code{AnalyticsMetricsBase$export_csv()}}
\item \href{#method-AnalyticsMetricsBase-dataframe}{\code{AnalyticsMetricsBase$dataframe()}}
\item \href{#method-AnalyticsMetricsBase-clear_cache}{\code{AnalyticsMetricsBase$clear_cache()}}
\item \href{#method-AnalyticsMetricsBase-format}{\code{AnalyticsMetricsBase$format()}}
\item \href{#method-AnalyticsMetricsBase-clone}{\code{AnalyticsMetricsBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalyticsMetricsBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnalyticsMetricsBase-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``AnalyticsMetricsBase`` instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalyticsMetricsBase$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalyticsMetricsBase-export_csv"></a>}}
\if{latex}{\out{\hypertarget{method-AnalyticsMetricsBase-export_csv}{}}}
\subsection{Method \code{export_csv()}}{
ersists the analytics dataframe to a file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalyticsMetricsBase$export_csv(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{(str): The name of the file to save to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalyticsMetricsBase-dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-AnalyticsMetricsBase-dataframe}{}}}
\subsection{Method \code{dataframe()}}{
A dataframe with lots of interesting results about this
             object. Created by calling SageMaker List and Describe APIs and
             converting them into a convenient tabular summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalyticsMetricsBase$dataframe(force_refresh = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force_refresh}}{(bool): Set to True to fetch the latest data from
SageMaker API.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalyticsMetricsBase-clear_cache"></a>}}
\if{latex}{\out{\hypertarget{method-AnalyticsMetricsBase-clear_cache}{}}}
\subsection{Method \code{clear_cache()}}{
Clear the object of all local caches of API methods, so that the next
             time any properties are accessed they will be refreshed from the
             service.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalyticsMetricsBase$clear_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalyticsMetricsBase-format"></a>}}
\if{latex}{\out{\hypertarget{method-AnalyticsMetricsBase-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalyticsMetricsBase$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnalyticsMetricsBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnalyticsMetricsBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnalyticsMetricsBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
