% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{deprecated_function}
\alias{deprecated_function}
\title{Wrap a function with a deprecation warning.}
\usage{
deprecated_function(func, name)
}
\arguments{
\item{func:}{Function to wrap in a deprecation warning.}

\item{name:}{The name that has been deprecated.}
}
\value{
The modified function
}
\description{
Wrap a function with a deprecation warning.
}
