% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{.invoke}
\alias{.invoke}
\title{Safe method to invoke paws functions}
\usage{
.invoke(.fn, args)
}
\description{
Safe method to invoke paws functions
}
\examples{
\dontrun{
  library(R6sagemaker.common)

  s3 <- paws::s3()

  params <- list(
     Bucket = "made-up"
  )
  .invoke(s3$list_objects_v2, params)
}
}
\keyword{internal}
