% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{FeatureStoreOutput}
\alias{FeatureStoreOutput}
\title{Amazon SageMaker Feature Store class}
\description{
Configuration for processing job outputs in Amazon SageMaker Feature Store
}
\section{Super class}{
\code{\link[sagemaker.core:ApiObject]{sagemaker.core::ApiObject}} -> \code{FeatureStoreOutput}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{feature_group_name}}{placeholder}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureStoreOutput-clone}{\code{FeatureStoreOutput$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="format"><a href='../../sagemaker.core/html/ApiObject.html#method-ApiObject-format'><code>sagemaker.core::ApiObject$format()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="from_paws"><a href='../../sagemaker.core/html/ApiObject.html#method-ApiObject-from_paws'><code>sagemaker.core::ApiObject$from_paws()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="initialize"><a href='../../sagemaker.core/html/ApiObject.html#method-ApiObject-initialize'><code>sagemaker.core::ApiObject$initialize()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="to_paws"><a href='../../sagemaker.core/html/ApiObject.html#method-ApiObject-to_paws'><code>sagemaker.core::ApiObject$to_paws()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureStoreOutput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureStoreOutput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureStoreOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
