% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger.R
\name{get_default_profiler_rule}
\alias{get_default_profiler_rule}
\title{Return the default built-in profiler rule with a unique name.}
\usage{
get_default_profiler_rule()
}
\value{
ProfilerRule: The instance of the built-in ProfilerRule.
}
\description{
Return the default built-in profiler rule with a unique name.
}
