% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger.R
\name{get_rule_container_image_uri}
\alias{get_rule_container_image_uri}
\title{Return the Debugger rule image URI for the given AWS Region.}
\usage{
get_rule_container_image_uri(region)
}
\arguments{
\item{region}{(str): A string of AWS Region. For example, ``'us-east-1'``.}
}
\value{
str : Formatted image URI for the given AWS Region and the rule container type.
}
\description{
For a full list of rule image URIs,
             see `Use Debugger Docker Images for Built-in or Custom Rules
             \url{https://docs.aws.amazon.com/sagemaker/latest/dg/debugger-docker-images-rules.html}.
}
