% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{is.dir}
\alias{is.dir}
\title{check if file location is a directory or not}
\usage{
is.dir(directory)
}
\arguments{
\item{directory}{(str): directory path location}
}
\description{
check if file location is a directory or not
}
\keyword{internal}
