% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_properties.R
\name{MetadataProperties}
\alias{MetadataProperties}
\title{Accepts metadata properties parameters for conversion to request dict.}
\description{
Accepts metadata properties parameters for conversion to request dict.

Accepts metadata properties parameters for conversion to request dict.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MetadataProperties$new()}}
\item \href{#method-to_request_list}{\code{MetadataProperties$to_request_list()}}
\item \href{#method-format}{\code{MetadataProperties$format()}}
\item \href{#method-clone}{\code{MetadataProperties$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``MetadataProperties`` instance and turn parameters into dict.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataProperties$new(
  commit_id = NULL,
  repository = NULL,
  generated_by = NULL,
  project_id = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commit_id}}{(str):}

\item{\code{repository}}{(str):}

\item{\code{generated_by}}{(str):}

\item{\code{project_id}}{(str):}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_request_list}{}}}
\subsection{Method \code{to_request_list()}}{
Generates a request dictionary using the parameters provided to the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataProperties$to_request_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataProperties$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetadataProperties$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
