% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarify.R
\name{PDPConfig}
\alias{PDPConfig}
\title{Config class for Partial Dependence Plots (PDP).}
\description{
If PDP is requested, the Partial Dependence Plots will be included in the report, and the
corresponding values will be included in the analysis output.
}
\section{Super class}{
\code{\link[sagemaker.common:ExplainabilityConfig]{sagemaker.common::ExplainabilityConfig}} -> \code{PDPConfig}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pdp_config}}{PDP Config}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PDPConfig$new()}}
\item \href{#method-get_explainability_config}{\code{PDPConfig$get_explainability_config()}}
\item \href{#method-clone}{\code{PDPConfig$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ExplainabilityConfig" data-id="format">}\href{../../sagemaker.common/html/ExplainabilityConfig.html#method-format}{\code{sagemaker.common::ExplainabilityConfig$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes config for PDP.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDPConfig$new(features = NULL, grid_resolution = 15, top_k_features = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{features}}{(None or list): List of features names or indices for which partial dependence
plots must be computed and plotted. When ShapConfig is provided, this parameter is
optional as Clarify will try to compute the partial dependence plots for top
feature based on SHAP attributions. When ShapConfig is not provided, 'features'
must be provided.}

\item{\code{grid_resolution}}{(int): In case of numerical features, this number represents that
number of buckets that range of values must be divided into. This decides the
granularity of the grid in which the PDP are plotted.}

\item{\code{top_k_features}}{(int): Set the number of top SHAP attributes to be selected to compute
partial dependence plots.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_explainability_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_explainability_config}{}}}
\subsection{Method \code{get_explainability_config()}}{
Returns config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDPConfig$get_explainability_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDPConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
