% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{ProcessingOutput}
\alias{ProcessingOutput}
\title{ProcessingOutput Class}
\description{
Accepts parameters that specify an Amazon S3 output for a processing job and provides
             a method to turn those parameters into a dictionary.
}
\seealso{
Other Processor: 
\code{\link{ProcessingInput}},
\code{\link{ProcessingJob}},
\code{\link{Processor}},
\code{\link{ScriptProcessor}}
}
\concept{Processor}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProcessingOutput-new}{\code{ProcessingOutput$new()}}
\item \href{#method-ProcessingOutput-to_request_list}{\code{ProcessingOutput$to_request_list()}}
\item \href{#method-ProcessingOutput-format}{\code{ProcessingOutput$format()}}
\item \href{#method-ProcessingOutput-clone}{\code{ProcessingOutput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessingOutput-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessingOutput-new}{}}}
\subsection{Method \code{new()}}{
Initializes a ``ProcessingOutput`` instance. ``ProcessingOutput`` accepts parameters that
             specify an Amazon S3 output for a processing job and provides a method to turn
             those parameters into a dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessingOutput$new(
  source = NULL,
  destination = NULL,
  output_name = NULL,
  s3_upload_mode = c("EndOfJob", "Continuous"),
  app_managed = FALSE,
  feature_store_output = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{(str): The source for the output.}

\item{\code{destination}}{(str): The destination of the output. If a destination
is not provided, one will be generated:
"s3://<default-bucket-name>/<job-name>/output/<output-name>".}

\item{\code{output_name}}{(str): The name of the output. If a name
is not provided, one will be generated (eg. "output-1").}

\item{\code{s3_upload_mode}}{(str): Valid options are "EndOfJob" or "Continuous".
s3_upload_mode (str): Valid options are "EndOfJob" or "Continuous".}

\item{\code{app_managed}}{(bool): Whether the input are managed by SageMaker or application}

\item{\code{feature_store_output}}{(:class:`~sagemaker.processing.FeatureStoreOutput`)
Configuration for processing job outputs of FeatureStore.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessingOutput-to_request_list"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessingOutput-to_request_list}{}}}
\subsection{Method \code{to_request_list()}}{
Generates a request dictionary using the parameters provided to the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessingOutput$to_request_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessingOutput-format"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessingOutput-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessingOutput$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessingOutput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessingOutput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessingOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
