% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{R6sagemaker.common-package}
\alias{R6sagemaker.common}
\alias{R6sagemaker.common-package}
\title{r6sagemaker.common: this is just a placeholder}
\description{
`R6sagemaker` lower level api calls.
}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
