% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecations.R
\name{renamed_kwargs}
\alias{renamed_kwargs}
\title{Checks if the deprecated argument is in kwargs}
\usage{
renamed_kwargs(old_name, new_name, value, kwargs)
}
\arguments{
\item{old_name:}{name of deprecated argument}

\item{new_name:}{name of the new argument}

\item{value:}{value associated with new name, if supplied}

\item{kwargs:}{keyword arguments dict}
}
\value{
value of the keyword argument, if present
}
\description{
Raises warning, if present.
}
