% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{RunArgs}
\alias{RunArgs}
\title{RunArgs Class}
\description{
Accepts parameters that correspond to ScriptProcessors.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{code}}{This can be an S3 URI or a local path to a file with the framework script to run}

\item{\code{inputs}}{Input files for the processing job}

\item{\code{outputs}}{Outputs for the processing job}

\item{\code{arguments}}{A list of string arguments to be passed to a processing job}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RunArgs-new}{\code{RunArgs$new()}}
\item \href{#method-RunArgs-clone}{\code{RunArgs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunArgs-new"></a>}}
\if{latex}{\out{\hypertarget{method-RunArgs-new}{}}}
\subsection{Method \code{new()}}{
An instance of this class is returned from the ``get_run_args()`` method on processors,
            and is used for normalizing the arguments so that they can be passed to
            :class:`~sagemaker.workflow.steps.ProcessingStep`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunArgs$new(code, inputs = NULL, outputs = NULL, arguments = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{(str): This can be an S3 URI or a local path to a file with the framework
script to run.}

\item{\code{inputs}}{(list[:class:`~sagemaker.processing.ProcessingInput`]): Input files for
the processing job. These must be provided as
:class:`~sagemaker.processing.ProcessingInput` objects (default: None).}

\item{\code{outputs}}{(list[:class:`~sagemaker.processing.ProcessingOutput`]): Outputs for
the processing job. These can be specified as either path strings or
:class:`~sagemaker.processing.ProcessingOutput` objects (default: None).}

\item{\code{arguments}}{(list[str]): A list of string arguments to be passed to a
processing job (default: None).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RunArgs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RunArgs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RunArgs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
