% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_definition_inputs.R
\name{S3Input}
\alias{S3Input}
\title{Metadata of data objects stored in S3.}
\description{
Two options are provided: specifying a S3 prefix or by explicitly listing the files
             in  manifest file and referencing the manifest file's S3 path.
}
\note{
Note: Strong consistency is not guaranteed if S3Prefix is provided here.
             S3 list operations are not strongly consistent.
             Use ManifestFile if strong consistency is required.
}
\section{Super class}{
\code{\link[sagemaker.core:ApiObject]{sagemaker.core::ApiObject}} -> \code{S3Input}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{S3Input$new()}}
\item \href{#method-clone}{\code{S3Input$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="format">}\href{../../sagemaker.core/html/ApiObject.html#method-format}{\code{sagemaker.core::ApiObject$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="from_paws">}\href{../../sagemaker.core/html/ApiObject.html#method-from_paws}{\code{sagemaker.core::ApiObject$from_paws()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="to_paws">}\href{../../sagemaker.core/html/ApiObject.html#method-to_paws}{\code{sagemaker.core::ApiObject$to_paws()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize S3Input.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Input$new(
  s3_uri = NULL,
  local_path = NULL,
  s3_data_type = "S3Prefix",
  s3_input_mode = "File",
  s3_data_distribution_type = "FullyReplicated",
  s3_compression_type = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s3_uri}}{(str, default=None): the path to a specific S3 object or a S3 prefix}

\item{\code{local_path}}{(str, default=None):
the path to a local directory. If not provided, skips data download
by SageMaker platform.}

\item{\code{s3_data_type}}{(str, default="S3Prefix"): Valid options are "ManifestFile" or "S3Prefix".}

\item{\code{s3_input_mode}}{(str, default="File"): Valid options are "Pipe" or "File".}

\item{\code{s3_data_distribution_type}}{(str, default="FullyReplicated"):
Valid options are "FullyReplicated" or "ShardedByS3Key".}

\item{\code{s3_compression_type}}{(str, default=None): Valid options are "None" or "Gzip"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Input$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
