% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger_metrics_config.R
\name{SMDataParallelProfilingConfig}
\alias{SMDataParallelProfilingConfig}
\title{SMDataParallelProfilingConfig Class}
\description{
Configuration for framework metrics collected from a SageMaker Distributed training job.
}
\section{Super class}{
\code{\link[sagemaker.common:MetricsConfigBase]{sagemaker.common::MetricsConfigBase}} -> \code{SMDataParallelProfilingConfig}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SMDataParallelProfilingConfig-new}{\code{SMDataParallelProfilingConfig$new()}}
\item \href{#method-SMDataParallelProfilingConfig-clone}{\code{SMDataParallelProfilingConfig$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.common" data-topic="MetricsConfigBase" data-id="format"><a href='../../sagemaker.common/html/MetricsConfigBase.html#method-MetricsConfigBase-format'><code>sagemaker.common::MetricsConfigBase$format()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.common" data-topic="MetricsConfigBase" data-id="to_json_string"><a href='../../sagemaker.common/html/MetricsConfigBase.html#method-MetricsConfigBase-to_json_string'><code>sagemaker.common::MetricsConfigBase$to_json_string()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMDataParallelProfilingConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-SMDataParallelProfilingConfig-new}{}}}
\subsection{Method \code{new()}}{
Specify target steps or a target duration to profile.
             By default, it profiles step 15 of training.
             If \code{profile_default_steps} is set to `True` and none of the other
             range parameters is specified,
             the class uses the default configuration for SageMaker Distributed profiling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMDataParallelProfilingConfig$new(
  start_step = NULL,
  num_steps = NULL,
  start_unix_time = NULL,
  duration = NULL,
  profile_default_steps = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_step}}{(int): The step to start profiling. The default is step 15.}

\item{\code{num_steps}}{(int): The number of steps to profile. The default is for 1 steps.}

\item{\code{start_unix_time}}{(int): The Unix time to start profiling.}

\item{\code{duration}}{(float): The duration in seconds to profile.}

\item{\code{profile_default_steps}}{(bool): Indicates whether the default configuration
should be used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SMDataParallelProfilingConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SMDataParallelProfilingConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SMDataParallelProfilingConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
