% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{TrainingJobAnalytics}
\alias{TrainingJobAnalytics}
\title{TrainingJobAnalytics Class}
\description{
Fetch training curve data from CloudWatch Metrics for a specific training
             job.
}
\section{Super class}{
\code{\link[sagemaker.common:AnalyticsMetricsBase]{sagemaker.common::AnalyticsMetricsBase}} -> \code{TrainingJobAnalytics}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{CLOUDWATCH_NAMESPACE}}{CloudWatch namespace to return Training Job Analytics data}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{Name of the TrainingJob being analyzed}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TrainingJobAnalytics-new}{\code{TrainingJobAnalytics$new()}}
\item \href{#method-TrainingJobAnalytics-clear_cache}{\code{TrainingJobAnalytics$clear_cache()}}
\item \href{#method-TrainingJobAnalytics-clone}{\code{TrainingJobAnalytics$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sagemaker.common" data-topic="AnalyticsMetricsBase" data-id="dataframe"><a href='../../sagemaker.common/html/AnalyticsMetricsBase.html#method-AnalyticsMetricsBase-dataframe'><code>sagemaker.common::AnalyticsMetricsBase$dataframe()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.common" data-topic="AnalyticsMetricsBase" data-id="export_csv"><a href='../../sagemaker.common/html/AnalyticsMetricsBase.html#method-AnalyticsMetricsBase-export_csv'><code>sagemaker.common::AnalyticsMetricsBase$export_csv()</code></a></li>
<li><span class="pkg-link" data-pkg="sagemaker.common" data-topic="AnalyticsMetricsBase" data-id="format"><a href='../../sagemaker.common/html/AnalyticsMetricsBase.html#method-AnalyticsMetricsBase-format'><code>sagemaker.common::AnalyticsMetricsBase$format()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainingJobAnalytics-new"></a>}}
\if{latex}{\out{\hypertarget{method-TrainingJobAnalytics-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``TrainingJobAnalytics`` instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingJobAnalytics$new(
  training_job_name,
  metric_names = NULL,
  sagemaker_session = NULL,
  start_time = NULL,
  end_time = NULL,
  period = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{training_job_name}}{(str): name of the TrainingJob to analyze.}

\item{\code{metric_names}}{(list, optional): string names of all the metrics to
collect for this training job. If not specified, then it will
use all metric names configured for this job.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other
AWS services needed. If not specified, one is specified using
the default AWS configuration chain.}

\item{\code{start_time}}{:}

\item{\code{end_time}}{:}

\item{\code{period}}{:}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainingJobAnalytics-clear_cache"></a>}}
\if{latex}{\out{\hypertarget{method-TrainingJobAnalytics-clear_cache}{}}}
\subsection{Method \code{clear_cache()}}{
Clear the object of all local caches of API methods, so that the next
             time any properties are accessed they will be refreshed from the
             service.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingJobAnalytics$clear_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrainingJobAnalytics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TrainingJobAnalytics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingJobAnalytics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
