% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_utils.R
\docType{data}
\name{vpc_config}
\alias{vpc_config}
\title{Vpc Configuration}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 4.
}
\usage{
vpc_config
}
\description{
Contains default vpc configurations.
}
\keyword{internal}
