% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_definition_inputs.R
\name{AthenaDatasetDefinition}
\alias{AthenaDatasetDefinition}
\title{DatasetDefinition for Athena.}
\description{
With this input, SQL queries will be executed using Athena to generate datasets to S3.
}
\section{Super class}{
\code{\link[sagemaker.core:ApiObject]{sagemaker.core::ApiObject}} -> \code{AthenaDatasetDefinition}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AthenaDatasetDefinition$new()}}
\item \href{#method-clone}{\code{AthenaDatasetDefinition$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="format">}\href{../../sagemaker.core/html/ApiObject.html#method-format}{\code{sagemaker.core::ApiObject$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="from_paws">}\href{../../sagemaker.core/html/ApiObject.html#method-from_paws}{\code{sagemaker.core::ApiObject$from_paws()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="to_paws">}\href{../../sagemaker.core/html/ApiObject.html#method-to_paws}{\code{sagemaker.core::ApiObject$to_paws()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize AthenaDatasetDefinition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AthenaDatasetDefinition$new(
  catalog = NULL,
  database = NULL,
  query_string = NULL,
  output_s3_uri = NULL,
  work_group = NULL,
  kms_key_id = NULL,
  output_format = NULL,
  output_compression = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalog}}{(str, default=None): The name of the data catalog used in Athena query
execution.}

\item{\code{database}}{(str, default=None): The name of the database used in the Athena query
execution.}

\item{\code{query_string}}{(str, default=None): The SQL query statements, to be executed.}

\item{\code{output_s3_uri}}{(str, default=None):
The location in Amazon S3 where Athena query results are stored.}

\item{\code{work_group}}{(str, default=None):
The name of the workgroup in which the Athena query is being started.}

\item{\code{kms_key_id}}{(str, default=None): The AWS Key Management Service (AWS KMS) key that Amazon
SageMaker uses to encrypt data generated from an Athena query execution.}

\item{\code{output_format}}{(str, default=None): The data storage format for Athena query results.
Valid options are "PARQUET", "ORC", "AVRO", "JSON", "TEXTFILE"}

\item{\code{output_compression}}{(str, default=None): The compression used for Athena query results.
Valid options are "GZIP", "SNAPPY", "ZLIB"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AthenaDatasetDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
