% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_definition_inputs.R
\name{DatasetDefinition}
\alias{DatasetDefinition}
\title{DatasetDefinition input.}
\description{
DatasetDefinition input.

DatasetDefinition input.
}
\section{Super class}{
\code{\link[sagemaker.core:ApiObject]{sagemaker.core::ApiObject}} -> \code{DatasetDefinition}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatasetDefinition$new()}}
\item \href{#method-clone}{\code{DatasetDefinition$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="format">}\href{../../sagemaker.core/html/ApiObject.html#method-format}{\code{sagemaker.core::ApiObject$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="from_paws">}\href{../../sagemaker.core/html/ApiObject.html#method-from_paws}{\code{sagemaker.core::ApiObject$from_paws()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.core" data-topic="ApiObject" data-id="to_paws">}\href{../../sagemaker.core/html/ApiObject.html#method-to_paws}{\code{sagemaker.core::ApiObject$to_paws()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize DatasetDefinition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDefinition$new(
  data_distribution_type = "ShardedByS3Key",
  input_mode = "File",
  local_path = NULL,
  redshift_dataset_definition = NULL,
  athena_dataset_definition = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_distribution_type}}{(str, default="ShardedByS3Key"):
Whether the generated dataset is FullyReplicated or ShardedByS3Key (default).}

\item{\code{input_mode}}{(str, default="File"):
Whether to use File or Pipe input mode. In File (default) mode, Amazon
SageMaker copies the data from the input source onto the local Amazon Elastic Block
Store (Amazon EBS) volumes before starting your training algorithm. This is the most
commonly used input mode. In Pipe mode, Amazon SageMaker streams input data from the
source directly to your algorithm without using the EBS volume.}

\item{\code{local_path}}{(str, default=None):
The local path where you want Amazon SageMaker to download the Dataset
Definition inputs to run a processing job. LocalPath is an absolute path to the
input data. This is a required parameter when `AppManaged` is False (default).}

\item{\code{redshift_dataset_definition}}{(:class:\code{sagemaker.common::RedshiftDatasetDefinition},
default=None):
Configuration for Redshift Dataset Definition input.}

\item{\code{athena_dataset_definition}}{(:class:\code{sagemaker.common::AthenaDatasetDefinition},
default=None):
Configuration for Athena Dataset Definition input.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatasetDefinition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
