% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarify.R
\name{TextConfig}
\alias{TextConfig}
\title{Config object to handle text features.}
\description{
The SHAP analysis will break down longer text into chunks (e.g. tokens, sentences, or paragraphs
             ) and replace them with the strings specified in the baseline for that feature. The shap value
             of a chunk then captures how much replacing it affects the prediction.
}
\section{Super class}{
\code{\link[sagemaker.common:ExplainabilityConfig]{sagemaker.common::ExplainabilityConfig}} -> \code{TextConfig}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{text_config}}{Text Config}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TextConfig$new()}}
\item \href{#method-get_text_config}{\code{TextConfig$get_text_config()}}
\item \href{#method-clone}{\code{TextConfig$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ExplainabilityConfig" data-id="format">}\href{../../sagemaker.common/html/ExplainabilityConfig.html#method-format}{\code{sagemaker.common::ExplainabilityConfig$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ExplainabilityConfig" data-id="get_explainability_config">}\href{../../sagemaker.common/html/ExplainabilityConfig.html#method-get_explainability_config}{\code{sagemaker.common::ExplainabilityConfig$get_explainability_config()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a text configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextConfig$new(granularity, language)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{granularity}}{(str): Determines the granularity in which text features are broken down
to, can be "token", "sentence", or "paragraph". Shap values are computed for these units.}

\item{\code{language}}{(str): Specifies the language of the text features, can be "chinese", "danish",
"dutch", "english", "french", "german", "greek", "italian", "japanese", "lithuanian",
"multi-language", "norwegian bokmål", "polish", "portuguese", "romanian", "russian",
"spanish", "afrikaans", "albanian", "arabic", "armenian", "basque", "bengali", "bulgarian",
"catalan", "croatian", "czech", "estonian", "finnish", "gujarati", "hebrew", "hindi",
"hungarian", "icelandic", "indonesian", "irish", "kannada", "kyrgyz", "latvian", "ligurian",
"luxembourgish", "macedonian", "malayalam", "marathi", "nepali", "persian", "sanskrit",
"serbian", "setswana", "sinhala", "slovak", "slovenian", "swedish", "tagalog", "tamil",
"tatar", "telugu", "thai", "turkish", "ukrainian", "urdu", "vietnamese", "yoruba". Use
"multi-language" for a mix of mulitple languages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_text_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_text_config}{}}}
\subsection{Method \code{get_text_config()}}{
Returns part of an analysis config dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextConfig$get_text_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
